Former-commit-id: a337bf2089ddaccea1ba64da7625d1e3be74e6e1

+++ acrg_hbmcmc/hbmcmc.py
@@ -191,9 +191,6 @@ def fixedbasisMCMC(species, sites, domain, meas_period, start_date,
             tempdir = basis.quadtreebasisfunction(emissions_name, fp_all, sites, 
                           start_date, domain, species, outputname,
                           nbasis=nbasis)
-            #             tempdir = quadtree.quadtreebasisfunction(emissions_name, fp_all, sites, 
-#                           start_date, domain, species, outputname,
-#                           nbasis=nbasis)
             fp_basis_case= "quadtree"+species+"-"+outputname
             basis_directory = tempdir
     else:
@@ -250,10 +247,11 @@ def fixedbasisMCMC(species, sites, domain, meas_period, start_date,
                            xprior, bcprior, sigprior,Ytime, siteindicator, sigma_freq_index, data, fp_data,
                            emissions_name, domain, species, sites,
                            start_date, end_date, outputname, outputpath,
-                           basis_directory, country_file, fp_basis_case, country_unit_prefix)
-    
-    # remove the temporary basis function directory
-    shutil.rmtree(tempdir)
+                           basis_directory, country_file, country_unit_prefix)
+
+    if quadtree_basis is True:
+        # remove the temporary basis function directory
+        shutil.rmtree(tempdir)
     
     print("All done")
